+++ DataAccess/Models/Offer.cs
@@ -14,10 +14,10 @@ namespace DataAccess.Models
         public string Id { get; set; }
         
         public string Title { get; set; }
-        public decimal Area { get; set; }
+        public double Area { get; set; }
         public string AreaUnit { get; set; }
-        public decimal PricePerUnit { get; set; }
-        public decimal Price { get; set; }
+        public double PricePerUnit { get; set; }
+        public double Price { get; set; }
 
         public string Url { get; set; }
         public string Location { get; set; }
